PeerFinder work and refactoring: * Implement PeerFinder business logic. * Support fixed peers (including DNS support). * Add journal support to Peer and Peers. * Refactor PeerDoor support. * Tidy up Peers and eliminate connection functionality and timers. * Refactor Peer interface and add journal support. * Allow construction of incoming Peer using an existing socket. * Remove TESTNET support. * Allow connections from/to cluster peers without consuming slots * Misc. cleanups.

+++ src/ripple_app/peers/PackedMessage.h
@@ -45,7 +45,7 @@ public:
 public:
     /** Number of bytes in a message header.
     */
-    static unsigned const kHeaderBytes = 6;
+    static size_t const kHeaderBytes = 6;
 
     PackedMessage (::google::protobuf::Message const& message, int type);
